requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
anthropic==0.97.0
stripe==15.1.0
requests==2.32.3
python-dotenv==1.2.2
pydantic==2.13.3
